// src/common/Design_Components/Colors/Color.types.ts
export enum Colors {
  Grey_20 = 'grey_20',
  Grey_40 = 'grey_40',
  Grey_60 = 'grey_60',
  Grey_80 = 'grey_80',
  Black = 'black',
  White = 'white',
  Transparent = 'transparent',
  Success = 'success',
  Error = 'error',
  Warning = 'warning',
  Primary = 'primary',
  Primary_80 = 'primary_80',
  Primary_60 = 'primary_60',
  Primary_40 = 'primary_40',
  Primary_20 = 'primary_20'
}